requirement.txt
fastapi==0.95.1
uvicorn==0.22.0
pinecone-client==2.0.0
langchain==0.0.147
langchain_pinecone==0.1.0
langchain-groq==0.0.1
huggingface-hub==0.13.3
python-dotenv==1.0.0
pydantic==1.10.6